backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
tortoise-orm==0.22.2
aerich==0.8.1
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.10.4
